Adds testing to Glucose Reading Form
Introduces component and integration tests for the GlucoseReadingForm.

Adds data-testid attributes to form elements to facilitate testing.
This allows for reliable identification of elements within the tests.

Adds tests for form validation and successful submission.
These tests help ensure the form functions as expected and provides useful feedback to the user.